Refactor: Se creo un hook personalizado y se agrego en el readme los framework usados  - [Maximiliano Calpanchay]
