+++ README.md
@@ -1,8 +1,15 @@
 # React + Vite
-## Integrantes del Grupo
+## 👥 Integrantes del Grupo
 
 - **Calpanchay Maximiliano** – **maxiicalpa**
 - **Padilla Santiago** – **santip019**
 - **Fernandez Cruz Bruno** – **brunofernandez877**
 - **Efimov Valentina** – **Valen-08**
 - **Cruz Antonella Guadalupe** - **antocruz06**
+
+## 🚀 Tecnologías Utilizadas en el proyecto
+
+- ⚡ **React** – Biblioteca principal para construir la interfaz de usuario.
+- ⚡ **Vite** – Herramienta de desarrollo rápido y bundler moderno.
+- ⚡ **Bootstrap** – Framework CSS para el diseño responsivo y componentes visuales.
+- ⚡ **React Router DOM** – Librería para la navegación y rutas dentro de la aplicación.
